requirements.txt
dj-database-url
Django
gunicorn
Pillow
whitenoise
django-widget-tweaks
psycopg2
django-localflavor